Check if kitty theme is already set before updating
The timer now runs every 60s, and changes the theme if its name is not
found in the `~/.config/kitty/current-theme.conf` file.

This prevents the case when resetting the same theme would reset the
dynamically adjusted font size of Kitty.  Furthermore, it enables the
timer to run independently of the actual time we want to set for when a
theme should be active.